Flag brew nvrs which are out of date compared to latest in brewweb.
For this to work, we need to add a cronjob that will do a GET to:

http://localhost:3000/packages/get_latest_pkgs_from_brew/-1?task_id=jb-eap-6-dot-2-dot-0&secret_key=birdistheword

Flagging will be done by highlighting the brew field yellow.

The huge disadvantage is that this process takes a lot of time and will block
any other requests made while the request above is serviced.