Cleanup ett a little bit
As ETT (and in the future, Wildbee) grows, we need to better manage the existing
code. My first attempt to cleanup ETT a little bit is by deleting the unused
helper functions that get autogenerated by Ruby on Rails, in the hopes that it
will create less confusion when people will look at the application. Why?
Because those helpers look like dead code to me and dead code leads to confusion
("Is it in use or not?")

My goal was inspired from this [blog][rails_spikes] post.

In the future, I'd like to add more unit tests and maybe also try to move code
into their rightful place. I've been guilty of putting code in the wrong areas
while I was learning Rails and I'd like to fix my mistakes.

As such, I've learned that helpers are meant to be really helpers for views and
nothing else. I'll therefore move the stuff in application_helper.rb to their
rightful place in a future commit.

My ultimate goal is to make the codebase easier to read, and maintain when ETT
will become legacy and replaced by its replacement, Wildbee.

Wildbee itself is not as future complete as ETT but given some time, it should
be, thanks to the efforts of Francisco and Miles.

[rails_spikes]: http://railspikes.com/2008/8/22/how-to-fix-your-rails-helpers