Remove any N+1 issue in the generation of packages index
This was done by simply observing the logs and seeing what ActiveRecord was
loading.

The `get_packages` method was rewritten, to stop using direct SQL statements but
also to make use of eager loading.

The method that was looking up information if a package is in shipped-list has
been disabled (`in_shipped_list?`)since it is too slow (took ~ 8s).