Redirect to correct path when login
Let's say you are not signed into ETT, and you are looking at the Tasks page.
You remind yourself that you need to sign-in to be able to edit a task; and so,
you click the 'Login' link.

The Login page, upon authentication, will just bring you to the front page
instead of the previously visited page (task).

This commit attempts to fix this inconvenience by redirecting the user to the
previously visited page upon being 'logged in'.