Add ability to send edited file contents to another server
When user will press 'Submit Build', the edited file contents will be sent to
the other server responsible for builds.

User will have option to not send the edited file contents to the other server
via a checkbox option.

When the user will move the package having the edited file contents to the
'Finished' state, the edited file contents will be deleted from ETT's database.