Add support for Github Pull Request
When someone marks a package as done, sometimes (s)he has to do a PR. This
commits add a feature where ETT will track PRs and see if they are marked as
close for you.

How does it work?

1. Create a token on Github to be able to access Github repositories and save it
   in an environmental variable 'GITHUB_ETT_TOKEN'

2. Set the github repo to track in the environmental variable 'GITHUB_REPO_LINK'

3. ???

4. Profit!

Everytime you do a get on '/toolbox/update_github_pr_information', ETT will
ask Github whether a PR is already marked as closed or not. You can then 'get'
this link in a cronjob if you want to ask Github about the status of a PR
regularly.

It will scan through active tasks, and search for packages that have unclosed
prs. It will then ask Github if they are still marked as closed or not.

A tick or a cross will be used to indicate if the PR is merged or not
respectively.

TODO:
Add a view in the packages list to see those PRs.