Remove frozen state and unify readonly tasks
Previously the ReadOnly task had a separate view from the general tasks
view. This commit unifies readonly with the tasks view so that you can
mark a task as 'readonly' from the task view. Fancy!

The frozen state was removed. Instead, a task is 'frozen' when it is not
'cronjob active'. I know. Confusing. Oh well...