Clarified some text regarding frozen. Also first rails4 commit!
Wait what? What Rails4 commit?

The commit on package.rb is an attempt to make sure the models are going
to work great on Rails4. This is my first attempt and we'll try to see
how it works. For now any Rails2 specific code will be delimited by 'if'
guards. Later on if the migration is successful the 'if' guards will be
removed.