ReadOnly task: don't touch pkgs of different prods
When a task is put on readonly, similar packages in different tasks will
get their nvrs, bugs removed from their respective erratas.

This behaviour is generally ok, but only if we touch the similar
packages being in the same prod as the origin task marked as readonly.
Otherwise we shouldn't do it.